agrisense/requirements.txt
streamlit
transformers
torch
requests
pillow
groq
python-dotenv
